# perf: cache x11 block header hash, reduce reindex hashes from 4->2 per block