--- a/src/primitives/block.cpp
+++ b/src/primitives/block.cpp
@@ -12,10 +12,13 @@
 
 uint256 CBlockHeader::GetHash() const
 {
-    std::vector<unsigned char> vch(80);
-    CVectorWriter ss(SER_GETHASH, PROTOCOL_VERSION, vch, 0);
-    ss << *this;
-    return HashX11((const char *)vch.data(), (const char *)vch.data() + vch.size());
+    if (cached_hash.IsNull()) {
+        std::vector<unsigned char> vch(80);
+        CVectorWriter ss(SER_GETHASH, PROTOCOL_VERSION, vch, 0);
+        ss << *this;
+        cached_hash = HashX11((const char *)vch.data(), (const char *)vch.data() + vch.size());
+    }
+    return cached_hash;
 }
 
 std::string CBlock::ToString() const