## Issue being fixed or feature implemented
We currently hash the header up to 4 times per block we are loading from disk; drop this down to 2 times

before:
./src/dashd --printtoconsole --testnet --nowallet --reindex --stopatheight=1   76.34s user 6.15s system 95% cpu 1:25.98 total
![image](https://github.com/user-attachments/assets/5368e0f8-0a61-416a-936a-fa3744807f4c)

after:

./src/dashd --printtoconsole --testnet --nowallet --reindex --stopatheight=1   62.87s user 5.70s system 95% cpu 1:11.52 total
![image](https://github.com/user-attachments/assets/e1e6438d-50b9-4320-b51d-4224898806d4)

Shaves ~20-25% off of header reindex times.
Calculated as ((76.34-62.87)/76.34)=17.6%, but this includes the overhead of startup / shutdown, so it's likely higher.

## What was done?


## How Has This Been Tested?
Reindex headers; should probably do a full reindex

## Breaking Changes

## Checklist:
  _Go over all the following points, and put an `x` in all the boxes that apply._
- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have added or updated relevant unit/integration/functional/e2e tests
- [ ] I have made corresponding changes to the documentation
- [x] I have assigned this pull request to a milestone _(for repository code-owners and collaborators only)_