@UdjinM6 2025-03-07
Nice! 👍 

Could also avoid resetting `cached_hash` on writes and squeeze a bit more out of it:
```diff
diff --git a/src/primitives/block.h b/src/primitives/block.h
index 158c42b2c1..d80fb30215 100644
--- a/src/primitives/block.h
+++ b/src/primitives/block.h
@@ -41,7 +41,7 @@ public:
 
     SERIALIZE_METHODS(CBlockHeader, obj) {
         READWRITE(obj.nVersion, obj.hashPrevBlock, obj.hashMerkleRoot, obj.nTime, obj.nBits, obj.nNonce);
-        obj.cached_hash.SetNull();
+        SER_READ(obj, obj.cached_hash.SetNull());
     }
 
     void SetNull()
```

develop:
119.50s user 13.16s system 98% cpu 2:14.32 total

This PR:
101.09s user 12.94s system 97% cpu 1:56.42 total

This PR with the patch above:
89.55s user 12.55s system 97% cpu 1:44.35 total